Distances5
Moved the get_distance function to processing.py and called it from there in categories.py. It now works well But i still need to merge the results into a shapefile of the urban areas to manually verify at least.
Also began in this commit to add del vars to try to free some space when variables are no longer used in the script.