Categories
Modified the fact that to process categories, you could only use an external raster you would've cropped earlier using qgis. Now, the function raster crop export the file then return the path of it so another function can reoppen it to work with it again. The reason i save the file and don't just return it as i woul do with a dataframe,is because the metadata of the image is updated appart from the file (output_meta.update...) and it can be merged only when saved (with open(......, **output_meta)).