Save
To intersect and crop every polygon from a layer, I create a generator with an input gdf then execute the main function with the resulting polygon which i need to save then reopen. In cause, the raster crop function which uses fiona to open the shapefile and not geopandas like everything else.