Urban extent
Added a dist_village variable in the config file with the goal of modifying the get_urban_extent() function to gather urban patches in the radius of x meters to be considered as the same village. A problem might ocur when a urban patches is already in a village but aligible for another one due to its distance.