Fix for the desynchronized indexes result in the geodataframe.
Up until this fix, the GeoDataFrame i got as a result desynchronized as new values weren't added inplace of the cells were they should be added. This error originated from an inplace drop of rows which left the index with holes. Then when i tried to iterate over it, the indexes mismatched whereas the number of rows was the same.