Redesign of the main file
Massive redesign of the main file to move from an approach where i used a function to intersect the data then process it to another approach where a single function holds the datasets paths + the various intersections + the functions to start the processings.
This redesign allow to to perform the same process on multiple data (pas, buffers, everything...). It also allow the addition of a top-level progress bar to track what is being processed at the time. Finally, it, as the previous function did, allow to process a choosen area of interest.