Fix for the faulty insertion of the Series of land use
After processing the land use proportionality, i pivoted the table to get the labels as the column names and the values still as values. But when inserting it into the GeoDataFrame, it generated an error or at best appended below as new lines.
The line 182 fix that.

I also renamed the function from get_pas_profiles to get_profile because the type of area which is processed depends on the GeoDataFrame given as an input thus it is not a PA bu default