Filepath, not dirpath and Modules not found but installed
First: I modified the dirpath function because i changed the approach and do not want to check for every shapefiles from a directory but rather want to check for every polygon from a shapefile.
Second: When executing from the CLI, python doesn't  look for the modules of the venv but checks for the modules of the os install instead.